# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VLTONE Platform API
  version: 1.0.0
  description: Source-of-truth contract shared by the public site, admin site, and desktop application.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/browser-backgrounds/{backgroundID}/image:
    get:
      operationId: getPublicBrowserBackgroundImage
      parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
      responses:
        "200": { description: Sanitized image; hidden images require admin session, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/browser-backgrounds/{backgroundID}/thumbnail:
    get:
      operationId: getPublicBrowserBackgroundThumbnail
      parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
      responses:
        "200": { description: Sanitized image; hidden images require admin session, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/admin/browser-backgrounds:
    get:
      operationId: listAdminBrowserBackgrounds
      security: [{ adminCookie: [] }]
      responses:
        "200": { description: Backgrounds ordered by sort_order then newest first; public list includes only published images, content: { application/json: { schema: { $ref: "#/components/schemas/BrowserBackgroundList" } } } }
    post:
      operationId: uploadBrowserBackground
      security: [{ adminCookie: [], csrf: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, file]
              properties:
                title: { type: string, minLength: 1, maxLength: 160 }
                file: { type: string, format: binary, description: PNG/JPEG/WebP up to 10 MiB and 16 MP; server re-encodes and generates a thumbnail }
      responses:
        "201": { description: Published background, content: { application/json: { schema: { $ref: "#/components/schemas/BrowserBackground" } } } }
        "422": { $ref: "#/components/responses/Error" }
...
  /v1/admin/browser-backgrounds/{backgroundID}/image:
    get:
      operationId: getAdminBrowserBackgroundImage
      security: [{ adminCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
      responses:
        "200": { description: Sanitized image; hidden images require admin session, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/admin/browser-backgrounds/{backgroundID}/thumbnail:
    get:
      operationId: getAdminBrowserBackgroundThumbnail
      security: [{ adminCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/BackgroundID" }]
      responses:
        "200": { description: Sanitized image; hidden images require admin session, content: { image/jpeg: { schema: { type: string, format: binary } }, image/png: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    webCookie: { type: apiKey, in: cookie, name: vlt_web_session }
    adminCookie: { type: apiKey, in: cookie, name: vlt_admin_session }
    csrf: { type: apiKey, in: header, name: X-CSRF-Token }
    desktopBearer: { type: http, scheme: bearer, bearerFormat: Ed25519 }
    reporterBearer: { type: http, scheme: bearer, bearerFormat: opaque }
  parameters:
    BackgroundID: { name: backgroundID, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Error:
      description: Structured API error
      content: { application/json: { schema: { $ref: "#/components/schemas/APIError" } } }
...
  schemas:
    BrowserBackgroundList:
      type: object
      required: [backgrounds]
      properties:
        backgrounds: { type: array, maxItems: 100, items: { $ref: "#/components/schemas/BrowserBackground" } }
    BrowserBackground:
      type: object
      required: [id, title, published, sort_order, mime_type, bytes, width, height, sha256, url, thumbnail_url, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, maxLength: 160 }
        published: { type: boolean }
        sort_order: { type: integer }
        mime_type: { type: string, enum: [image/png, image/jpeg] }
        bytes: { type: integer, maximum: 20971520 }
        width: { type: integer }
        height: { type: integer }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        url: { type: string }
        thumbnail_url: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    APIError:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        field_errors: { type: object, additionalProperties: { type: string } }
        request_id: { type: string }
...